+++ packages/components/src/organisms/ProductShelf/ProductShelfItems.tsx
@@ -1,0 +1,24 @@
+import type { HTMLAttributes } from 'react'
+import React, { forwardRef } from 'react'
+
+export interface ProductShelfItemsProps extends HTMLAttributes<HTMLUListElement> {
+  /**
+   * ID to find this component in testing tools (e.g.: Cypress, Testing Library, and Jest).
+   */
+  testId?: string
+}
+
+const ProductShelfItems = forwardRef<HTMLUListElement, ProductShelfItemsProps>(
+  function ProductShelfItems(
+    { testId = 'fs-product-shelf-items', children, ...otherProps },
+    ref
+  ) {
+    return (
+      <ul ref={ref} data-fs-product-shelf-items data-testid={testId} {...otherProps}>
+        {children}
+      </ul>
+    )
+  }
+)
+
+export default ProductShelfItems